Update summarize.py
I tried to implement the "output language selector" raised (by me :D ) in #3 .

As this is my absolute first commit/PR ever, I am more than thankful for any feedback on what to change or improve :)

The script now takes an additional optional parameter: "language"
Default is still english, so without parameter the current querys are untouched.
if language is specified, at the end of the querys a sentence is added that output should be in specified language.

In my tests, "german" or "deutscher" as parameter has better results than "deutsch"....